rename header in atoms  to Typography
