--- src/components/atoms/themes.ts
@@ -1,9 +1,0 @@
-export const theme = {
-  colors: {
-    header: "#000000",
-    subheader: "#808080",
-    body: "#000000",
-    footer: "#000000",
-  },
-  mobile: "768px",
-};
